test(pagination): Refactor visual testing stories to use CSF
